Added scanner objects to app.java

+++ Simple Calculator/src/app/ArithmeticOperations.java
@@ -1,19 +1,19 @@
 package app;
 
 public class ArithmeticOperations {
-    public int add(int x, int y) {
+    public double add(double x, double y) {
         return x + y;
     }
 
-    public int sub(int x, int y) {
+    public double sub(double x, double y) {
         return x - y;
     }
 
-    public int divide(int x, int y) {
+    public double divide(double x, double y) {
         return x / y;
     }
 
-    public int multiply(int x, int y) {
+    public double multiply(double x, double y) {
         return x * y;
     }
 }
